refactor: query/ 패키지를 책임별 서브패키지로 분리 (#97)
10개 클래스가 혼재된 query/ 패키지를 4개 서브패키지로 분리:
- template/ : QueryTemplate, QueryTemplateRegistry
- validation/ : SqlValidator, QueryParameterValidator
- execution/ : SqlExecutor
- generation/ : DynamicSqlGenerator, ParameterExtractor
