+++ src/main/java/com/flab/woowahaneats/domain/chatbot/query/generation/DynamicSqlGenerator.java
@@ -1,4 +1,4 @@
-package com.flab.woowahaneats.domain.chatbot.query;
+package com.flab.woowahaneats.domain.chatbot.query.generation;
 
 import com.flab.woowahaneats.domain.chatbot.llm.LlmExecutor;
 import com.flab.woowahaneats.domain.chatbot.prompt.PromptProvider;
