+++ src/main/java/com/flab/woowahaneats/domain/chatbot/query/validation/SqlValidator.java
@@ -1,4 +1,4 @@
-package com.flab.woowahaneats.domain.chatbot.query;
+package com.flab.woowahaneats.domain.chatbot.query.validation;
 
 import lombok.extern.slf4j.Slf4j;
 import net.sf.jsqlparser.parser.CCJSqlParserUtil;
